pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py[voice]>=2.5.2",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pynacl>=1.5.0",
    "python-dotenv>=1.1.0",
    "spotdl>=4.2.11",
    "tlgbotfwk>=0.4.61",
    "yt-dlp>=2025.3.25",
]
